radicle: OS specific setting of file limits
On Windows systems the file limits are of type `i64` in comparison to `u64` on Unix
systems.

In order to attempt both, the `set_file_limit` function is duplicated to handle
the case for i64 for Windows, and u64 for Unix. Each is guarded by its
respective `target_family`.

In order to support the `usize` configuration value, the functions are modified
to accept values that can fallibly be converted to `i64` and `u64`,
respectively. If this fails an `io::Error` is returned.

X-Clacks-Overhead: GNU Terry Pratchett

+++ radicle-node/src/main.rs
@@ -120,7 +120,7 @@ fn execute() -> anyhow::Result<()> {
         config.node.listen.clone()
     };
 
-    if let Err(e) = radicle::io::set_file_limit(config.node.limits.max_open_files as u64) {
+    if let Err(e) = radicle::io::set_file_limit(config.node.limits.max_open_files) {
         log::warn!(target: "node", "Unable to set process open file limit: {e}");
     }
